delete button added / separate handling

+++ src/Controller/SettingsController.php
@@ -569,41 +569,45 @@ class SettingsController extends Controller
         }
 
         if ($form->isSubmitted() && $form->isValid()) {
-            // send invitation email
-            if (isset($data['email'])) {
-                $invitationCode = $invitationsService->generateInvitationCode($authSourceItem, $roomId, $data['email']);
+            $clickedButton = $form->getClickedButton()->getName();
 
-                $invitationLink = $request->getSchemeAndHttpHost();
-                $invitationLink .= '?cid=' . $portal->getItemId() . '&mod=home&fct=index&cs_modus=portalmember';
-                $invitationLink .= '&invitation_auth_source=' . $authSourceItem->getItemId();
-                $invitationLink .= '&invitation_auth_code=' . $invitationCode;
+            if ($clickedButton === 'send') {
+                // send invitation email
+                if (isset($data['email'])) {
+                    $invitationCode = $invitationsService->generateInvitationCode($authSourceItem, $roomId, $data['email']);
 
-                $mailer = $this->get('mailer');
-                $fromAddress = $this->getParameter('commsy.email.from');
-                $fromSender = $legacyEnvironment->getCurrentContextItem()->getContextItem()->getTitle();
+                    $invitationLink = $request->getSchemeAndHttpHost();
+                    $invitationLink .= '?cid=' . $portal->getItemId() . '&mod=home&fct=index&cs_modus=portalmember';
+                    $invitationLink .= '&invitation_auth_source=' . $authSourceItem->getItemId();
+                    $invitationLink .= '&invitation_auth_code=' . $invitationCode;
 
-                $subject = $translator->trans('invitation subject %portal%', [
-                    '%portal%' => $portal->getTitle(),
-                ]);
-                $body = $translator->trans('invitation body %portal% %link% %sender%', [
-                    '%room%' => $roomItem->getTitle(),
-                    '%portal%' => $portal->getTitle(),
-                    '%link%' => $invitationLink,
-                    '%roomLink%' => $this->generateUrl('app_room_home', [
-                        'roomId' => $roomItem->getItemID(),
-                    ], UrlGeneratorInterface::ABSOLUTE_URL),
-                    '%sender%' => $user->getFullName(),
-                ]);
-                $mailMessage = (new \Swift_Message())
-                    ->setSubject($subject)
-                    ->setBody($body, 'text/plain')
-                    ->setFrom([$fromAddress => $fromSender])
-                    ->setTo([$data['email']]);
-                $mailer->send($mailMessage);
-            }
+                    $mailer = $this->get('mailer');
+                    $fromAddress = $this->getParameter('commsy.email.from');
+                    $fromSender = $legacyEnvironment->getCurrentContextItem()->getContextItem()->getTitle();
 
-            foreach ($data['remove_invitees'] as $removeInvitee) {
-                $invitationsService->removeInvitedEmailAdresses($authSourceItem, $removeInvitee);
+                    $subject = $translator->trans('invitation subject %portal%', [
+                        '%portal%' => $portal->getTitle(),
+                    ]);
+                    $body = $translator->trans('invitation body %portal% %link% %sender%', [
+                        '%room%' => $roomItem->getTitle(),
+                        '%portal%' => $portal->getTitle(),
+                        '%link%' => $invitationLink,
+                        '%roomLink%' => $this->generateUrl('app_room_home', [
+                            'roomId' => $roomItem->getItemID(),
+                        ], UrlGeneratorInterface::ABSOLUTE_URL),
+                        '%sender%' => $user->getFullName(),
+                    ]);
+                    $mailMessage = (new \Swift_Message())
+                        ->setSubject($subject)
+                        ->setBody($body, 'text/plain')
+                        ->setFrom([$fromAddress => $fromSender])
+                        ->setTo([$data['email']]);
+                    $mailer->send($mailMessage);
+                }
+            } else if ($clickedButton === 'delete') {
+                foreach ($data['remove_invitees'] as $removeInvitee) {
+                    $invitationsService->removeInvitedEmailAdresses($authSourceItem, $removeInvitee);
+                }
             }
 
             return $this->redirectToRoute('app_settings_invitations', [
